migrate : cleanup-export-string to export-string
