--- packages/cleanup-export-string/errors/index.ts
@@ -1,1 +1,0 @@
-export * from "./no-token-interpreter";
