--- packages/export-string/coli-stringfy/declarations/index.ts
@@ -1,3 +1,0 @@
-export * from "./variable";
-export * from "./import";
-export * from "./function";
