--- packages/export-string/coli-stringfy/expressions/index.ts
@@ -1,1 +1,0 @@
-export * from "./comment";
