--- packages/export-string/coli-stringfy/index.ts
@@ -1,2 +1,0 @@
-export * from "./expressions";
-export * from "./declarations";
