+++ packages/export-string/core/declarations/function.ts
