+++ packages/export-string/core/declarations/index.ts
