+++ packages/export-string/core/declarations/variable.ts
