--- packages/export-string/package.json
@@ -1,4 +1,0 @@
-{
-    "name": "@coli/export-string",
-    "version": "0.0.0"
-}
